fix: cast long positionId to int before Postgrest .Filter calls
Supabase.Postgrest.Table<T>.Filter<TCriterion>(...) only accepts
string, int, float, List, Dictionary<string, object>,
FullTextSearchConfig and Range criterion types - not long. Three
repository methods (RunRepository.GetLatestPayloadHashAsync,
RankRepository.GetExistingRanksAsync, RankRepository.GetTopRanksAsync)
passed the long-typed positionId straight into .Filter("position_id",
...), which throws PostgrestException: Unknown criterion type against
the real client.

Found by running the write-mode job against production: every
position failed identically at this call before any write occurred.
Invisible to the existing suite since repositories are mocked in
tests, never exercised against the real Postgrest client.

Cast positionId to int at the three call sites (position ids are
always 1-5, so this is safe) rather than switching to the .Where(lambda)
pattern, to keep the fix minimal and scoped to the type mismatch.

Fixes #20

+++ Services/Repositories/RankRepository.cs
@@ -29,7 +29,9 @@ public class RankRepository : IRankRepository
             .From<FantasyProsRank>()
             .Filter("season", Constants.Operator.Equals, season)
             .Filter("week", Constants.Operator.Equals, week)
-            .Filter("position_id", Constants.Operator.Equals, positionId)
+            // Postgrest's Filter(...) only accepts string/int/float/List/Dictionary/FullTextSearchConfig/Range
+            // criterion types (not long) - cast down; position ids are always small (1-5).
+            .Filter("position_id", Constants.Operator.Equals, (int)positionId)
             .Filter("scoring", Constants.Operator.Equals, scoring)
             .Get(cancellationToken);
 
@@ -47,7 +49,9 @@ public class RankRepository : IRankRepository
             .From<FantasyProsRank>()
             .Filter("season", Constants.Operator.Equals, season)
             .Filter("week", Constants.Operator.Equals, week)
-            .Filter("position_id", Constants.Operator.Equals, positionId)
+            // Postgrest's Filter(...) only accepts string/int/float/List/Dictionary/FullTextSearchConfig/Range
+            // criterion types (not long) - cast down; position ids are always small (1-5).
+            .Filter("position_id", Constants.Operator.Equals, (int)positionId)
             .Filter("scoring", Constants.Operator.Equals, scoring)
             .Order("rank_ecr", Constants.Ordering.Ascending)
             .Limit(limit)
